fix: move delivery details serializer to be nested with order serializer
